[kube-prometheus-stack] Fix NetworkPolicy for Alertmanager
Set gateway/ingress namespace and podLabel selector to an empty string/object in order to not accidentally expose something unexpected. Move the Loki rule to the additionalIngress examples as Loki is not part of the stack. The 'alertmanager.service.clusterPort' was added as it was not declared in the values and now needs to be enabled explicitly.

Issue: #5318
Signed-off-by: Sebastian Gaiser <sebastian.gaiser@hetzner-cloud.de>